requirements.txt
transformers
torch
psycopg2-binary
typer
python-dotenv
sqlalchemy
pandas
